% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataGeneration.R
\name{datagen}
\alias{datagen}
\title{Function to generate simulation data}
\usage{
datagen(n = 1000, p = 300, joint_rank = 30,
  individual_rank = c(100, 120), nrep = 20, std = 0.1)
}
\arguments{
\item{n}{The number of rows of matrices that will be generated}

\item{p}{The number of columns of matrices that will be generated}

\item{joint_rank}{The true joint rank of matrices to be generated}

\item{individual_rank}{A vector of length two. The true individual ranks of matrices that will be generated}

\item{nrep}{Number of datasets to be generated. A pair of matrices is called a dataset.}

\item{std}{The standard deviation of normal distribution used to generatd noise matrices. Default is 0.1}
}
\value{
A list that contains the following:
\item{X1_list}{A list of first matrices generated.}
\item{X2_list}{A list of second matrices generated}
\item{trueJ1_list}{A list of true joint signal matrices of the \code{X1_list}}
\item{trueJ2_list}{A list of true joint signal matrices of the \code{X2_list}}
\item{trueI1_list}{A list of true individual signal matrices of the \code{X1_list}}
\item{trueI2_list}{A list of true individual signal matrices of the \code{X2_list}}
}
\description{
Function that generates a dataset using provided true joint rank and individual ranks. Number of matrices in one dataset is restricted to two.
}
\examples{
datagen(n = 100, p = 20, joint_rank = 5, individual_rank = c(5,7), nrep = 20)

}
