\name{DMMD-package}
\alias{DMMD-package}
\alias{DMMD}
\docType{package}
\title{
\packageTitle{DMMD}
}
\description{
\packageDescription{DMMD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{DMMD}
\packageIndices{DMMD}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{DMMD}

Maintainer: \packageMaintainer{DMMD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
