% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMD.R
\name{DMMD_Fit}
\alias{DMMD_Fit}
\title{Main function of double-matched matrix decomposition.}
\usage{
DMMD_Fit(
  X1,
  X2,
  r1 = NULL,
  r2 = NULL,
  rc = NULL,
  rr = NULL,
  angle_threshold = 90 * pi/180,
  variance1 = c("equal", "unequal"),
  variance2 = c("equal", "unequal"),
  method = c("PL", "ED"),
  tol = .Machine$double.eps^0.5,
  maxiter = 1000
)
}
\arguments{
\item{X1}{The first noisy data matrix}

\item{X2}{The second noisy data matrix}

\item{r1}{Total rank for X1. Default is NULL, meaning unknown, which will be estimated by rank estimation procedure determined by 'method'}

\item{r2}{Total rank for X2. Default is NULL, meaning unknown, which will be estimated by rank estimation procedure determined by 'method'}

\item{rc}{Joint column rank. Default is NULL, meaning unknown, which will be estimated by profile likelihood method}

\item{rr}{Joint row rank. Default is NULL, meaning unknown, which will be estimated by profile likelihood method}

\item{angle_threshold}{The threshold angle for principal angles. Principal angles greater than the threshold are not considered as joint signal. Default is 90 degree}

\item{variance1}{Either "equal" or "unequal". Default is "equal". This argument is the variance assumption used in the profile likelihood method for determining the total rank}

\item{variance2}{Either "equal" or "unequal". Default is "equal". This argument is the variance assumption used in the profile likelihood method for determining the joint rank}

\item{method}{The method used for determining the total ranks r1 and r2. Either "PL" (profile likelihood) or "ED" (edge distribution). Default is "PL"}

\item{tol}{The tolerance used to determine convergence. Default is the square root of the machine precision}

\item{maxiter}{Maximum number of iterations allowed in the iterative algorithm. Default is 1000}
}
\value{
A list with the following elements:
\item{r1}{The estimated total rank for X1, if not specified}
\item{r2}{The estimated total rank for X2, if not specified}
\item{rc}{The estimated joint column rank for X1, if not specified}
\item{rr}{The estimated joint row rank for X1, if not specified}
\item{A1}{The estimated low-rank signal matrix of X1}
\item{A2}{The estimated low-rank signal matrix of X2}
\item{E1}{The noise matrix of X1, \deqn{X_1 = A_1 + E_1}}
\item{E2}{The noise matrix of X2, \deqn{X_2 = A_2 + E_2}}
\item{Jc1}{The estimated low-rank joint column signal for X1}
\item{Jc2}{The estimated low-rank joint column signal for X2}
\item{Jr1}{The estimated low-rank joint row signal for X1}
\item{Jr2}{The estimated low-rank joint row signal for X2}
\item{Ic1}{The estimated low-rank individual column signal for X1}
\item{Ic2}{The estimated low-rank individual column signal for X2}
\item{Ir1}{The estimated low-rank individual row signal for X1}
\item{Ir2}{The estimated low-rank individual row signal for X2}
}
\description{
Main function of double-matched matrix decomposition.
}
\details{
This function decomposed double-matched matrices according to Lemma 1 in Dongbang Yuan & Irina Gaynanova (2022) Double-Matched Matrix Decomposition for Multi-View Data, Journal of Computational and Graphical Statistics, DOI: 10.1080/10618600.2022.2067860.
}
\examples{
data = DoubleDataGen(n = 20, p = 16, rank = c(4, 3), rc = 2, rr = 1, nrep = 1)
X1 = data$X1[[1]]
X2 = data$X2[[1]]
result_DMMD = DMMD_Fit(X1,X2)
}
