% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMD_Iterative.R
\name{DMMD_i}
\alias{DMMD_i}
\title{Main function of iterative DMMD algorithm}
\usage{
DMMD_i(
  X1,
  X2,
  r1 = NULL,
  r2 = NULL,
  rc = NULL,
  rr = NULL,
  angle_threshold = 90 * pi/180,
  variance1 = "equal",
  variance2 = "equal",
  method = "PL",
  eps = .Machine$double.eps^0.5,
  kmax = 1000,
  verbose = FALSE
)
}
\arguments{
\item{X1}{The first matrix.}

\item{X2}{The second matrix.}

\item{r1}{The total rank X1. Default is NULL, which means unknown.}

\item{r2}{The total rank X2. Default is NULL, which means unknown.}

\item{rc}{The joint column rank. Default is NULL, which means unknown.}

\item{rr}{The joint row rank. Default is NULL, which means unknown.}

\item{eps}{Tolerance, default is the square root of machine precision.}

\item{kmax}{The maximum iterations. Default is 1000.}

\item{verbose}{Do you want to see the progress of the function? Default is F, which means there is no progress shown.}
}
\description{
Main function of iterative DMMD algorithm
}
