% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOptMatrix.R
\name{FindOpt_DM_Iterative}
\alias{FindOpt_DM_Iterative}
\title{Function that iteratively calculates the optimal low-rank signal matrix given partial column and row space basis.}
\usage{
FindOpt_DM_Iterative(
  X,
  M,
  N,
  r,
  maxiter = 10000,
  tol = .Machine$double.eps^0.5
)
}
\arguments{
\item{X}{The noisy matrix. The dimension is n by p}

\item{M}{The column space that the signal matrix contains. The dimension is n by \deqn{r_c}}

\item{N}{The row space that the signal matrix contains. The dimension is p by \deqn{r_r}}

\item{r}{The rank of output signal matrix}

\item{maxiter}{The maximum number of iterations allowed in the calculation. Default is 1e4}

\item{tol}{The tolerance used to monitor the convergence. Default is the square root of machine precision}
}
\value{
A list with the following elements:
\item{result}{A matrix that is the estimated solution to the optimization problem}
\item{R}{The remaining \deqn{r - r_c} column bases}
\item{S}{The remaining \deqn{r - r_r} row bases}
\item{num_iter}{The actual number of iteration run}
\item{opt_function_val}{A vector of objective values calculated at each iteration. It is supposed to be monotonically decreasing}
}
\description{
Function that iteratively calculates the optimal low-rank signal matrix given partial column and row space basis.
}
\details{
This function solves the optimization problem of (2) in Dongbang Yuan & Irina Gaynanova (2022) Double-Matched Matrix Decomposition for Multi-View Data, Journal of Computational and Graphical Statistics, DOI: 10.1080/10618600.2022.2067860, that is:
\deqn{min_{A}{||X - A||^2_F}}
subject to \deqn{M is a subset of column space of A, N is a subset of row space of A}
}
\examples{
X = matrix(c(2,1,1,3,2,2),nrow = 3)
M = matrix(c(1,1,0), nrow = 3)
N = matrix(c(0,1), nrow = 2)
r = 2
FindOpt_DM_Iterative(X, M, N, r)
}
