% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOptMatrix.R
\name{FindOpt_SimplifiedCase}
\alias{FindOpt_SimplifiedCase}
\title{This function only works when either the rank of M or N equal to the specified rank r (\deqn{r = max(r_c, r_r)}). It's a prerequisite of FindOpt_DM_Iterative}
\usage{
FindOpt_SimplifiedCase(X, M, N, r)
}
\arguments{
\item{X}{The original matrix (m by n)}

\item{M}{The column space that the output matrix should have}

\item{N}{The row space that the output matrix should have}

\item{r}{The rank of output matrix}
}
\value{
A matrix that is the solution to the optimization problem.
}
\description{
This function only works when either the rank of M or N equal to the specified rank r (\deqn{r = max(r_c, r_r)}). It's a prerequisite of FindOpt_DM_Iterative
}
\details{
The resulting matrix is the solution the following problem:
\deqn{min_{Y}{||X - Y||^2_F}}
subject to \deqn{M is a subset of column space of Y, N is a subset of row space of Y}
}
