% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOptMatrix.R
\name{FindOpt_SM}
\alias{FindOpt_SM}
\title{Function that calculates the optimal low-rank signal matrix given partial column space basis}
\usage{
FindOpt_SM(X, M, r)
}
\arguments{
\item{X}{The noisy data matrix}

\item{M}{The column space that the output signal matrix contains}

\item{r}{The rank of output signal matrix}
}
\value{
A matrix that is the solution to the optimization problem.
}
\description{
Function that calculates the optimal low-rank signal matrix given partial column space basis
}
\details{
The resulting matrix is the solution the following problem:
\deqn{min_{Y}{||X - Y||^2_F}}
subject to \deqn{M is a subset of column space of Y}
}
\examples{
X = matrix(c(2,1,1,3,2,2),nrow = 3)
M = matrix(c(1,1,0),nrow = 3)
r = 2
FindOpt_SM(X,M,r)
}
