% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindOptMatrix.R
\name{FindOptMatrix}
\alias{FindOptMatrix}
\title{Function that calculates the optima matrix}
\usage{
FindOptMatrix(X, M, r)
}
\arguments{
\item{X}{The original matrix}

\item{M}{The matrix that contains}

\item{r}{The rank of output matrix}
}
\value{
A matrix that is the solution to the optimization problem.
}
\description{
Function that calculates the optima matrix
}
\details{
The resulting matrix is the solution the following problem:
\deqn{min_{Y}{||X - Y||^2_F}}
subject to \deqn{M is a subset of column space of Y}
}
\examples{
X = matrix(c(2,1,1,3,2,2),nrow = 3)
M = matrix(c(1,1,0),nrow = 3)
r = 2
FindOptMatrix(X,M,r)
}
