% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_Selection.R
\name{ProfileLikCluster}
\alias{ProfileLikCluster}
\title{Function that separates a vector into two groups using profile likelihood}
\usage{
ProfileLikCluster(x, variance = "equal")
}
\arguments{
\item{x}{Vector to be separated}

\item{variance}{Either "equal" or "unequal", corresponding to equal or unequal variance assumption. Default is "equal".}
}
\value{
A list that contains the following elements:
\item{index}{The index where separates the vector \code{x}}
\item{profileloglikvec}{A vector with profile log-likelihood for each index.}
}
\description{
Function that separates a vector into two groups using profile likelihood
}
\examples{
x = c(20,9.5,9,8,6,5,4.5,3)
ProfileLikCluster(x)
}
