% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_Selection.R
\name{ProfileLikClusterEqual}
\alias{ProfileLikClusterEqual}
\title{Preliminary function that separates a vector into two groups using profile likelihood assuming equal variance.}
\usage{
ProfileLikClusterEqual(x)
}
\arguments{
\item{x}{An ordered (max to min) vector to be clustered into two groups}
}
\value{
A list that contains the following elements:
\item{index}{The index that separates the vector \code{x}. The element at the index is considered in the first cluster}
\item{profileloglikvec}{A vector containing profile log-likelihood for each index}
}
\description{
Preliminary function that separates a vector into two groups using profile likelihood assuming equal variance.
}
\examples{
x = c(20,9.5,9,8,6,5,4.5,3)
ProfileLikClusterEqual(x)
}
