% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_Selection.R
\name{ProfileLoglik}
\alias{ProfileLoglik}
\title{Calculate the profile log-likelihood of given two data vectors assuming normal distributions.}
\usage{
ProfileLoglik(x, y, variance = c("equal", "unequal"))
}
\arguments{
\item{x}{The first data vector}

\item{y}{The second data vector}

\item{variance}{Either "equal" or "unequal", i.e., whether the assumption is equal variance or unequal variance. If no input is given, equal variance is assumed.}
}
\value{
The calculated profile log-likelihood
}
\description{
Calculate the profile log-likelihood of given two data vectors assuming normal distributions.
}
