% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Select_ED_Rank.R
\name{Select_ED_Rank}
\alias{Select_ED_Rank}
\title{Function that uses edge distribution (ED) method to calculate the rank i.e., the number of factors.}
\usage{
Select_ED_Rank(x, k_max = NULL, k_min = NULL, maxiter = 100)
}
\arguments{
\item{x}{The input vector for selecting the rank}

\item{k_max}{The maximum rank to be considered, restricted to less than or equal to half of the length of the vector. If not given, one tenth of the length of the vector is considered in general}

\item{k_min}{The minimum rank to be considered. If not given, it is treated as 1}

\item{maxiter}{The maximum iterations for the algorithm. Default 100}
}
\value{
The rank estimated by ED method
}
\description{
This function is a direct translation from Python function 'select_K_by_ED' in dcca.py.
}
\examples{
x = c(20,9.5,9,8,6,5,4.5,3)
Select_ED_Rank(x)
}
\references{
Onatski, Alexei. “DETERMINING THE NUMBER OF FACTORS FROM EMPIRICAL DISTRIBUTION OF EIGENVALUES.” The Review of Economics and Statistics 92, no. 4 (2010): 1004–16. http://www.jstor.org/stable/40985808.
}
