% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMMD.R
\name{SMMD}
\alias{SMMD}
\title{Single Matched Matrix Decomposition}
\usage{
SMMD(X1, X2, angle_threshold = 90 * pi/180, variance1 = "unequal",
  variance2 = "unequal", tol = .Machine$double.eps^0.5)
}
\arguments{
\item{X1}{The first matrix}

\item{X2}{The second matrix}

\item{angle_threshold}{Optional threshold (radians) feeded to \code{joint_angle_cluster} function. Default is 90 degrees, i.e., no threshold is used.}

\item{variance1}{Either "equal" or "unequal". The assumption of equal or unequal variance used to estimate the total rank.}

\item{variance2}{Either "equal" or "unequal". The assumption of equal or unequal variance used to estimate the joint rank.}

\item{tol}{Tolerence, default is the square root of machine precision.}
}
\value{
A list that contains the following:
\item{r1}{The estimated total rank of \code{X1}}
\item{r2}{The estimated total rank of \code{X2}}
\item{joint_rank}{The estimated rank of joint signals between \code{X1} and \code{X2}}
\item{individual_rank}{A vector of length 2. The vector contains the estimated ranks of individual signals of \code{X1} and \code{X2}}
\item{J1}{The joint signal of \code{X1}}
\item{J2}{The joint signal of \code{X2}}
\item{I1}{The individual signal of \code{X1}}
\item{I2}{The individual signal of \code{X2}}
}
\description{
Function that does single-matched matrix factorization. The matching is in the number of rows.
}
\details{
The decomposition is based on the theory in the paper of
Feng, Qing, et al. "Angle-based joint and individual variation explained." Journal of multivariate analysis 166 (2018): 241-265.
One can also refer to 
Lock, Eric F., et al. "Joint and individual variation explained (JIVE) for integrated analysis of multiple data types." The annals of applied statistics 7.1 (2013): 523.
for details.
}
\examples{
data = datagen(n = 100, p = 20, joint_rank = 5, individual_rank = c(5,7), nrep = 1)
X1 = data$X1_list[[1]]
X2 = data$X2_list[[1]]
result_SMMD = SMMD(X1,X2)
result_SMMD$joint_rank
result_SMMD$individual_rank

}
