% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angle_Calculation.R
\name{SUM_PCA_joint}
\alias{SUM_PCA_joint}
\title{Function that estimates joint row and column space by SUM-PCA.
X is the concatenated matrices (matched by rows)}
\usage{
SUM_PCA_joint(X, k = 2, joint_rank = NULL)
}
\description{
Function that estimates joint row and column space by SUM-PCA.
X is the concatenated matrices (matched by rows)
}
