% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_Selection.R
\name{VarianceMLE}
\alias{VarianceMLE}
\title{Calculate the MLE of variance of two vectors.}
\usage{
VarianceMLE(x, y, variance = c("equal", "unequal"))
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{variance}{Either "equal" or "unequal", i.e., whether the assumption is equal variance or unequal variance. If no input is given, equal variance is assumed.}
}
\value{
A vector of length 2 that contains the estimated variance for the two vectors
}
\description{
Calculate the MLE of variance of two vectors.
}
