% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angle_Calculation.R
\name{angle_cal}
\alias{angle_cal}
\title{Function that calculates principal angles between column spaces of two matrices}
\usage{
angle_cal(X, Y)
}
\arguments{
\item{X}{The first matrix.}

\item{Y}{The second matrix.}
}
\value{
A list that contains the following:
\item{angle}{A vector of principal angles with increasing order.}
\item{cos_angle}{A vector of cosine principal angles}
\item{principal_vector1}{Principal vectors of matrix \code{X}}
\item{principal_vector2}{Principal vectors of matrix \code{Y}}
}
\description{
Function that calculates principal angles between column spaces of two matrices
}
\examples{
X = matrix(c(1,1,1,1,1,0),nrow = 3, ncol = 2)
Y = matrix(c(1,1,1,2,1,0),nrow = 3, ncol = 2)
angle_cal(X,Y)

}
