% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myjivefunction.R
\name{col_jive}
\alias{col_jive}
\title{Calculate exact column joint and individual structure of two matrices}
\usage{
col_jive(A, B, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{A}{First matrix}

\item{B}{Second matrix}

\item{tol}{Tolerence, default is the square root of machine precision.}
}
\value{
A list that contains joint and individual signals. 
\item{J1}{Joint signal of first matrix}
\item{J2}{Joint signal of second matrix}
\item{I1}{Individual signal of the first matrix}
\item{I2}{Individual signal of the second matrix}
}
\description{
Detailed description...What is this?
}
\examples{
No example is a good example...
}
