% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Generation.R
\name{DoubleDataGen}
\alias{DoubleDataGen}
\title{Function that generates double-matched data matrices.}
\usage{
DoubleDataGen(
  n = 20,
  p = 16,
  rank = c(4, 3),
  rc = 2,
  rr = 1,
  nrep = 1,
  std1 = 0.1,
  std2 = 0.1,
  lb = 0.5,
  ub = 1.5
)
}
\arguments{
\item{n}{Number of rows of data matrix to be generated}

\item{p}{Number of rows of data matrix to be generated}

\item{rank}{The vector of total ranks of the two matrices to be generated}

\item{rc}{The joint column rank}

\item{rr}{The joint row rank}

\item{nrep}{How many pairs of double-matched matrices that you want to generate?}

\item{std1}{The standard deviation of the first Gaussian noise matrix}

\item{std2}{The standard deviation of the second Gaussian noise matrix}

\item{lb}{The lower bound for the singular values. Generated singular values are further scaled so that their sum of square equals to the total rank.}

\item{ub}{The upper bound for the singular values. Generated singular values are further scaled so that their sum of square equals to the total rank.}
}
\description{
Function that generates double-matched data matrices.
}
\examples{
data = DoubleDataGen()
head(data$X1[[1]])
}
