% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DoubleStandardize_function.R
\name{DoubleStandardize}
\alias{DoubleStandardize}
\title{Double standardize a matrix}
\usage{
DoubleStandardize(X, tol = .Machine$double.eps^0.5, maxIter = 100)
}
\arguments{
\item{X}{Matrix to be standardized.}

\item{tol}{Tolerance. Default is square root of machine precision.}

\item{maxIter}{Maximum iteration. Default is 100.}
}
\value{
A list that contains:
\item{Result}{The matrix after double standardization.}
\item{Iter}{Number of iterations the function actually runs.}
}
\description{
Double standardize a matrix
}
