% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Angle_Calculation.R
\name{joint_angle_cluster}
\alias{joint_angle_cluster}
\title{Function that estimates joint rank by method of profile likelihood.}
\usage{
joint_angle_cluster(
  angle_vec,
  angle_threshold = 90 * pi/180,
  variance = c("equal", "unequal")
)
}
\arguments{
\item{angle_vec}{A vector of principal angles}

\item{angle_threshold}{Optional threshold (radians) that is used to truncate principal angles. Default is 90 degrees, which means all the principal angles are taken into account.}

\item{variance}{Either "equal" or "unequal", i.e., whether the assumption is equal variance or unequal variance. If no input is given, equal variance is assumed.}
}
\value{
A list with the following elements:
\item{joint_rank}{Estimated joint rank}
\item{profileloglikvec}{Profile log likelihood calculated at each index. Function returns NA with a warning message, if less or equal to 2 principal angles are smaller than the threshold.}
}
\description{
Function that estimates joint rank by method of profile likelihood.
}
