% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_Selection.R
\name{ProfileLikCluster}
\alias{ProfileLikCluster}
\title{Function that separates a vector into two groups using profile likelihood}
\usage{
ProfileLikCluster(x, variance = c("equal", "unequal"))
}
\arguments{
\item{x}{An ordered (max to min) vector to be clustered into two groups}

\item{variance}{Either "equal" or "unequal", i.e., whether the assumption is equal variance or unequal variance. If no input is given, equal variance is assumed.}
}
\value{
A list that contains the following elements:
\item{index}{The index where separates the vector \code{x}. The element at the index is considered in the first cluster}
\item{profileloglikvec}{A vector containing profile log-likelihood for each index}
}
\description{
Function that separates a vector into two groups using profile likelihood
}
\examples{
x = c(20,9.5,9,8,6,5,4.5,3)
ProfileLikCluster(x)
}
