% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Profile_Likelihood_Rank_selection.R
\name{VarianceMLE}
\alias{VarianceMLE}
\title{Function that calculates MLE of variance of two vectors.}
\usage{
VarianceMLE(x, y, variance = "equal")
}
\arguments{
\item{x}{The first vector}

\item{y}{The second vector}

\item{variance}{Either "equal" or "unequal", 
corresponding to equal or unequal variance assumption.
Default is "equal".}
}
\value{
A vector of length two. If assumes equal variance, the result is the MLE version of pooled standard deviation repeating twice. If assumes unequal variance, the result is two MLE version of sample standard deviation.
}
\description{
Function that calculates MLE of variance of two vectors.
}
\examples{
x = c(1,2,3)
y = c(1,3,4)
VarianceMLE(x,y,variance = "equal")
VarianceMLE(x,y,variance = "unequal")

}
